Exploration on impact types, business models & selection criteria.
Initial research to understand and classify ecotech Categories based on the Drawdown project, types of positive and negative impact that can be evaluatd, business models for platform use, selection criteria for filtering out greenwashing and value-misaligned companies. Includes also a list of potential companies to invite for the pilot phase.
Created in December, 2024.